backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
Pillow==10.2.0
pydantic==2.6.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
httpx==0.26.0
pycairo==1.25.1
fonttools==4.47.2
